Add API-endpoint for updating filter.
Needs refinement. Need to make a copy of the default filter when a
change is made, so that it can stay as a reference. Maybe we need a
state for the filter on a page-level? Unsure about updating the DB on
every action, a save n seconds after last actions would maybe be
better.