Add duplication of default filter instead of editing it in place.
Added a state for the current filter in index, and pass it to every
component which will be updating the DB, maybe use Context instead, or
only have a state for the ID of the filter?
Returning new filter ID from requests, currently it's just jammed in the
response object, there is probably a nicer way.