Add weaponsection.
Added weaponsection, with options for the three archetypes currently
in-game. Refactored ButtonRow to be able to split itself into several
rows if needed (probably needs a rename). Refactored TierColumn to also
allow it to be used as a row, also needs to be renamed. Added new icons
for the weapon bases.
Currently there is a bug with the state of the options for the weapons,
the physical/elemental/tier buttons needs to update the state in the
weaponsection too, or not be overwritten during base selection.